feat: 튜닝 리포트 공개용 배치 기능 추가
